Add some new tests for various issues hooks/env/scalars
