+++ pkg/ddevapp/config_merge.go
@@ -1,7 +1,6 @@
 package ddevapp
 
 import (
-	"fmt"
 	"github.com/imdario/mergo"
 )
 
@@ -18,11 +17,6 @@ func (app *DdevApp) mergeAdditionalConfigIntoApp(configPath string) error {
 	err = mergo.Merge(app, newConfig, mergo.WithAppendSlice, mergo.WithOverride)
 	if err != nil {
 		return err
-	}
-
-	err = app.ValidateConfig()
-	if err != nil {
-		return fmt.Errorf("project config with file %s cannot be validated: %v", configPath, err)
 	}
 
 	return nil
