+++ pkg/ddevapp/config_test.go
@@ -12,7 +12,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/drud/ddev/pkg/ddevapp"
 	"github.com/drud/ddev/pkg/dockerutil"
 	"github.com/drud/ddev/pkg/exec"
 	"github.com/drud/ddev/pkg/globalconfig"
@@ -27,35 +26,6 @@ import (
 	"github.com/google/uuid"
 	asrt "github.com/stretchr/testify/assert"
 )
-
-// setupTestTempDir creates the test directory and related objects.
-func setupTestTempDir(t *testing.T) (*DdevApp, *asrt.Assertions) {
-	assert := asrt.New(t)
-	pwd, _ := os.Getwd()
-
-	projDir, err := filepath.Abs(testcommon.CreateTmpDir(t.Name()))
-	require.NoError(t, err)
-
-	testData := filepath.Join("./testdata/", t.Name(), "/.ddev")
-	err = fileutil.CopyDir(testData, filepath.Join(projDir, ".ddev"))
-	require.NoError(t, err)
-
-	app, err := NewApp(projDir, true)
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		err = os.Chdir(pwd)
-		assert.NoError(err)
-		err = app.Stop(true, false)
-		assert.NoError(err)
-		err = os.RemoveAll(projDir)
-		assert.NoError(err)
-	})
-	err = os.Chdir(app.AppRoot)
-	assert.NoError(err)
-
-	return app, assert
-}
 
 // TestNewConfig tests functionality around creating a new config, writing it to disk, and reading the resulting config.
 func TestNewConfig(t *testing.T) {
@@ -1371,135 +1341,4 @@ func TestDatabaseConfigUpgrade(t *testing.T) {
 		assert.Empty(app.MySQLVersion)
 		assert.Empty(app.MariaDBVersion)
 	}
-}
-
-// TestConfigMergeStringList verifies that scalar string merges update w/o clobber
-func TestConfigMergeStringList(t *testing.T) {
-	app, assert := setupTestTempDir(t)
-
-	// test matches allowing for delete syntax (a prefixed !)
-	assertSimpleMatch := func(expected bool, match string, setting []string) {
-		test := assert.True
-		if !expected {
-			test = assert.False
-		}
-
-		for _, val := range setting {
-			if val == match {
-				test(true, match)
-				return
-			}
-		}
-		test(false, match)
-	}
-	// no clobber of old setting
-	assertSimpleMatch(true, "somename", app.AdditionalHostnames)
-	// successful merge
-	assertSimpleMatch(true, "somename-new", app.AdditionalHostnames)
-}
-
-// TestConfigMergeEnvItems verifies that config overrides web_environment
-// and do not destroy (but may update) config.yaml stuff
-func TestConfigMergeEnvItems(t *testing.T) {
-	app, assert := setupTestTempDir(t)
-
-	// check the config file w/o overrides
-	noOverridesApp, err := ddevapp.NewApp(app.AppRoot, false)
-	require.NoError(t, err)
-	assert.IsType([]string{}, noOverridesApp.WebEnvironment)
-
-	// The app loaded without overrides should get the original values expected here
-	for _, v := range []string{`LARRY=l`, `MOE=m`, `CURLEY=c`} {
-		assert.Contains(noOverridesApp.WebEnvironment, v, "the app without overrides should have had %v but it didn't, webEnvironment=%v", v, noOverridesApp.WebEnvironment)
-	}
-
-	// With overrides we should have different values with the config.override.yaml values added
-	withOverridesApp, err := ddevapp.NewApp(app.AppRoot, true)
-	require.NoError(t, err)
-
-	for _, v := range []string{`LARRY=l`, `LARRY=lz`, `MOE=m`, `MOE=mz`, `CURLEY=c`, `SHEMP=s`} {
-		assert.Contains(withOverridesApp.WebEnvironment, v, "the app without overrides should have had %v but it didn't, webEnvironment=%v", v, noOverridesApp.WebEnvironment)
-	}
-
-	assert.Less(
-		nodeps.PosString(withOverridesApp.WebEnvironment, `LARRY=l`),
-		nodeps.PosString(withOverridesApp.WebEnvironment, `LARRY=lz`),
-		"LARRY=lz should come after LARRY=l",
-	)
-	assert.Less(
-		nodeps.PosString(withOverridesApp.WebEnvironment, `MOE=m`),
-		nodeps.PosString(withOverridesApp.WebEnvironment, `MOE=mz`),
-		"MOE=mz should come after MOE=m",
-	)
-}
-
-// TestConfigHooksMerge makes sure that hooks get merged with additional config.*.yaml
-func TestConfigHooksMerge(t *testing.T) {
-	app, _ := setupTestTempDir(t)
-
-	// some helpers
-	getHookTasks := func(hooks map[string][]ddevapp.YAMLTask, hook string) []ddevapp.YAMLTask {
-		tasks, ok := hooks[hook]
-		if !ok {
-			return nil
-		}
-		return tasks
-	}
-
-	// the fields we want are private!
-	hasTask := func(hooks map[string][]ddevapp.YAMLTask, hook, taskKey, desc string) bool {
-		tasks := getHookTasks(hooks, hook)
-		if tasks == nil {
-			return false
-		}
-		found := false
-		for _, task := range tasks {
-			taskDesc := ""
-			taskInterface, ok := task[taskKey]
-			if !ok {
-				// we guessed the key wrong
-				continue
-			}
-
-			taskDesc, ok = taskInterface.(string)
-			if !ok {
-				// we expected the command to be a string, but WTF?
-				continue
-			}
-
-			//t.Logf("key %s as %s", taskKey, taskDesc)
-
-			if taskDesc == desc {
-				found = true
-			}
-
-		}
-		return found
-	}
-
-	assertTask := func(expected bool, hook, taskKey, desc string) {
-		tasks := getHookTasks(app.Hooks, hook)
-		if tasks == nil {
-			if expected {
-				t.Errorf("did not found tasks for %s", hook)
-			} else {
-				return
-			}
-		} else {
-			found := hasTask(app.Hooks, hook, taskKey, desc)
-			if found != expected {
-				msg := "Expected "
-				if !expected {
-					msg = "Did not expect "
-				}
-				t.Errorf("%s Hook %s with %s", msg, hook, desc)
-			}
-		}
-	}
-
-	assertTask(true, "post-start", "exec", "simple random expression")
-	assertTask(true, "post-start", "exec-host", "simple host command")
-	assertTask(true, "post-start", "exec", "simple web command")
-	assertTask(true, "post-import-db", "exec", "drush uli")
-
 }
