+++ pkg/ddevapp/testdata/TestConfigMerge/envtest/expected/.ddev/config.yaml
@@ -1,0 +1,7 @@
+name: TestConfigMerget1
+web_environment:
+  - ORIG1=orig1
+  - ORIG_TO_BE_OVERWRITTEN=orig
+  - LAST=last_1
+  - LAST2=last_2
+  - ORIG_TO_BE_OVERWRITTEN=last
