+++ pkg/ddevapp/testdata/TestConfigMerge/hookstest/components/.ddev/config.last.yaml
@@ -1,0 +1,6 @@
+
+hooks:
+  pre-start:
+  - exec-host: "echo last pre-start"
+  post-start:
+  - exec: "echo last post-start"
