+++ pkg/ddevapp/testdata/TestConfigMerge/scalartest/components/.ddev/config.last.yaml
@@ -1,0 +1,1 @@
+webserver_type: apache-fpm
