+++ pkg/ddevapp/testdata/TestConfigMerge/scalartest/expected/.ddev/config.yaml
@@ -1,0 +1,8 @@
+
+database:
+  type: mariadb
+  version: 10.5
+mutagen_enabled: true
+webserver_type: apache-fpm
+
+php_version: 8.1
